Deleted check for empty variable **$version**
Checking that the variable **$version** is empty is completely redundant and limits functionality of helm. Because there are now way to use latest chart in repo (not for local) if `--version` parameter is set.

Docs:
```
--version string
specify the exact chart version to install. If this is not specified, the latest version is installed
```