Variable "helm_history_max" is now really working (#26)
* Variable "helm_history_max" is now really working (it before was parsed but not used)

- Appended **$history_max** variable to **$upgrade_args** (Note: if use `--helm_history_max N` instead `--helm_history_max=N` there is shell be error like `Error: unknown flag: --history-max N`).
- Changed arguments order in helm command to correspond docs (`helm upgrade [RELEASE] [CHART] [flags]`)
- Small text error

* Deleted check for empty variable **$version**

Checking that the variable **$version** is empty is completely redundant and limits functionality of helm. Because there are now way to use latest chart in repo (not for local) if `--version` parameter is set.

Docs:
```
--version string
specify the exact chart version to install. If this is not specified, the latest version is installed
```